backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.44.0",
    "fastapi>=0.116.1",
    "google-genai==1.11.0",
    "google-generativeai==0.8.5",
    "huggingface-hub[hf-xet]>=0.34.4",
    "instructor==1.10.0",
    "jinja2>=3.1.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
